+++ scripts/apps/vocabularies/components/VocabularyItemsViewEdit.tsx
